mcp-server/requirements.txt
fastmcp
starlette
pyyaml